% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{add_analysis}
\alias{add_analysis}
\title{Add Analysis}
\usage{
add_analysis(study, func = "list", params = list(), code = NULL,
  id = NULL)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{func}{The name of the function to run}

\item{params}{A list of parameters for the function arguments}

\item{code}{Code to define custom functions}

\item{id}{The id for this analysis (index or character) if NULL, this creates a new analysis, if an analysis with this id already exists, it will overwrite it}
}
\value{
A study object with class reg_study
}
\description{
Add an analysis to a study object
}
