% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_assumption.R
\name{add_assumption}
\alias{add_assumption}
\title{Add Assumption}
\usage{
add_assumption(
  study,
  id = NULL,
  description = "",
  value = NULL,
  hypothesis_id = NULL
)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for this assumption (index or character) if an assumption with this id already exists, it will overwrite it}

\item{description}{The text description of the assumption}

\item{value}{Whether the assumption is TRUE, FALSE or NULL (unknown)}

\item{hypothesis_id}{A vector of hypothesis IDs that rely on this assumption}
}
\value{
A study object with class scivrs_study
}
\description{
Add an auxillary assumption
}
\examples{

s <- study() \%>\%
  add_hypothesis("H1", "Speed and distance will be positively correlated") \%>\%
  add_assumption("normal_dist", "Distance will be normally distributed", "H1")

}
