% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_author.R
\name{add_author}
\alias{add_author}
\title{Add an author}
\usage{
add_author(study, surname, given = "", orcid = NULL, roles = c(), ...)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{surname}{a character string with the author's last name(s)}

\item{given}{a character string with the author's given name(s)}

\item{orcid}{the author's unique ORCiD (see https://orcid.org/)}

\item{roles}{a vector of roles from the CRediT taxonomy (see https://casrai.org/credit/); use credit_roles() to view the full list}

\item{...}{further info to add to author object}
}
\value{
A study object with class scivrs_study
}
\description{
Add an author
}
\examples{
s <- study() \%>\% add_author(
  surname = "DeBruine", given = "Lisa M.",
  orcid = "0000-0002-7523-5539",
  roles = c("Conceptualization", "Methodology")
)
study_to_json(s)

}
