% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{add_criterion}
\alias{add_criterion}
\title{Add Criterion}
\usage{
add_criterion(study, result, operator, comparator, hypothesis_id = NULL,
  analysis_id = NULL)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{result}{The name of the item in the analysis results list to compare}

\item{operator}{The operator for comparison c("<", "=", ">", "!=")}

\item{comparator}{The value to compare}

\item{hypothesis_id}{The id for the hypothesis (index or character) if NULL, assigns to the last hypothesis in the list}

\item{analysis_id}{The id for the relevant analysis (index or character) if NULL, assigns to the last analysis in the list}
}
\value{
A study object with class reg_study
}
\description{
Add a criterion to a hypothesis in a study object
}
\examples{

mystudy <- study("Iris Petals") \%>\%
  add_hypothesis("Petal length and width will be significantly correlated") \%>\%
  add_criterion("p.value", "<", 0.05)

}
