% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{add_hypothesis}
\alias{add_hypothesis}
\title{Add Hypothesis}
\usage{
add_hypothesis(study, id = NULL,
  description = "Describe your hypothesis", evaluation = "&")
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{id}{The id for this hypothesis (index or character) if NULL, this creates a new hypothesis, if a hypothesis with this id already exists, it will overwrite it}

\item{description}{The text descriiption of the hypothesis}

\item{evaluation}{The rule for evaluating this criterion c("&", "and", "|", "or")}
}
\value{
A study object with class reg_study
}
\description{
Add a hypothesis to a study object
}
\examples{

mystudy <- study("Iris Petals") \%>\%
  add_hypothesis(1, "Petal length and width will be significantly correlated")

}
