% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_method.R
\name{add_method}
\alias{add_method}
\title{Add a method}
\usage{
add_method(study, id = NULL, description = "Describe your method", ...)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for this method (index or character) if a method with this id already exists, it will overwrite it}

\item{description}{The text description of the method}

\item{...}{further arguments to add}
}
\value{
A study object with class scivrs_study
}
\description{
Add a method to a study object
}
\examples{

s <- study() \%>\%
  add_method("M1", "Description...")
study_to_json(s)

}
