% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{add_prep}
\alias{add_prep}
\title{Add Data Prep}
\usage{
add_prep(study, code, params = list(data = ".data[1]"),
  return = "data")
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{code}{Code to define the custom prep function or a reference to the file with the code}

\item{params}{A list of parameters for the function arguments}

\item{return}{A list of object names to return from the prep code}
}
\value{
A study object with class reg_study
}
\description{
Add a data prep step to a study object
}
