% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sim_data.R
\name{add_sim_data}
\alias{add_sim_data}
\title{Add simulated data}
\usage{
add_sim_data(
  study,
  data_id,
  within = list(),
  between = list(),
  n = 100,
  mu = 0,
  sd = 1,
  r = 0,
  empirical = FALSE,
  long = FALSE,
  dv = list(y = "value"),
  id = list(id = "id"),
  plot = FALSE,
  rep = 1
)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{data_id}{The id for this dataset (index or character) if a dataset with this id already exists, it will overwrite it}

\item{within}{a list of the within-subject factors}

\item{between}{a list of the between-subject factors}

\item{n}{the number of samples required}

\item{mu}{the means of the variables}

\item{sd}{the standard deviations of the variables}

\item{r}{the correlations among the variables (can be a single number, full correlation matrix as a matric or vector, or a vector of the upper right triangle of the correlation matrix}

\item{empirical}{logical. If true, mu, sd and r specify the empirical not population mean, sd and covariance}

\item{long}{Whether the returned tbl is in wide (default = FALSE) or long (TRUE) format}

\item{dv}{the name of the dv for long plots (defaults to y)}

\item{id}{the name of the id column (defaults to id)}

\item{plot}{whether to show a plot of the design}

\item{rep}{the number of data frames to return (default 1); if greater than 1, the returned data frame is nested by rep}
}
\value{
A study object with class scivrs_study
}
\description{
Uses \code{faux::sim_design()} to generate a data table with a specified within and between design.
}
\examples{

s <- study() \%>\%
  add_sim_data("dat", within = list(time = c("day", "night")),
               between = list(pet = c("cat", "dog")),
               n = 10, mu = 100, sd = 10, r = 0.5)
s$data[[1]]$data

}
