% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit.R
\name{author}
\alias{author}
\title{Create Author}
\usage{
author(name = c(), roles = c(), ...)
}
\arguments{
\item{name}{a named vector c(surname = "LastName", give = "First Names")}

\item{roles}{a vector of roles (e.g., c("Conceptualization", "Methodology")}

\item{...}{further info to add to author object}
}
\value{
a list with class scivrs_author
}
\description{
Create Author
}
\examples{
author(name = list(surname = "DeBruine", given = "Lisa M."),
       roles = list("Conceptualization", "Methodology"))
}
