% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit.R
\name{author_jats}
\alias{author_jats}
\title{Convert scivrs_author list to CRediT JATS format}
\usage{
author_jats(author)
}
\arguments{
\item{author}{list with class scivrs_author}
}
\value{
string in JATS format
}
\description{
Convert scivrs_author list to CRediT JATS format
}
\examples{
ld <- author(
  name = list(surname = "DeBruine", given = "Lisa M."),
  roles = list("Conceptualization", "Methodology")
)
author_jats(ld)
}
