% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_script.R
\name{cb_vars}
\alias{cb_vars}
\title{List codebook variables}
\usage{
cb_vars(cb)
}
\arguments{
\item{cb}{The codebook}
}
\value{
named list of all the variable descriptions
}
\description{
List codebook variables
}
\keyword{internal}
