% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit.R
\name{credit_roles}
\alias{credit_roles}
\title{CRediT Roles}
\usage{
credit_roles(display = c("names", "explain", "abbr"))
}
\arguments{
\item{display}{Whether to display the category names, explanations, or abbreviations}
}
\value{
list of roles
}
\description{
CRediT Roles
}
\examples{
credit_roles()
}
