% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{data_prep}
\alias{data_prep}
\title{Run data prep}
\usage{
data_prep(study)
}
\arguments{
\item{study}{A study list object with class reg_study}
}
\value{
A study object with class reg_study
}
\description{
Run the data prep on the raw data
}
