% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_eval.R
\name{eval_summary}
\alias{eval_summary}
\title{Evaluation summary}
\usage{
eval_summary(study)
}
\arguments{
\item{study}{A study list object with class scivrs_study}
}
\value{
Summary text
}
\description{
Evaluation summary
}
