% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_id}
\alias{fix_id}
\title{Fix IDs}
\usage{
fix_id(id)
}
\arguments{
\item{id}{the id to fix}
}
\value{
a fixed ID character string (only a-z, A-Z, 0-9, and _)
}
\description{
Fix IDs
}
\keyword{internal}
