% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{format_output}
\alias{format_output}
\title{Output as md, html or text}
\usage{
format_output(txt, output = c("md", "html", "text"))
}
\arguments{
\item{txt}{the md output of output_**** functions}

\item{output}{whether the output should be markdown, html, or plain text (defaults to md)}
}
\value{
character string in the specified format
}
\description{
Output as md, html or text
}
\keyword{internal}
