% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_from_xml.R
\name{full_text_table_from_grobid}
\alias{full_text_table_from_grobid}
\title{Get full text table from grobid}
\usage{
full_text_table_from_grobid(body)
}
\arguments{
\item{body}{body section of grobid xml as list (e.g., \code{xml2::as_list(xml)$TEI$text$body})}
}
\value{
a data frame of the classified full text
}
\description{
Get full text table from grobid
}
\keyword{internal}
