% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codebook.R
\name{get_codebook}
\alias{get_codebook}
\title{Get Codebook}
\usage{
get_codebook(study, data_id = 1)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{data_id}{The id for the dataset (index or character)}
}
\value{
a list with class scivrs_codebook
}
\description{
Get Codebook
}
\examples{
s <- study() \%>\% add_data("test", data.frame(x = 1:10))
cb <- get_codebook(s, "test")

}
