% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from study}
\usage{
get_data(study, id = 1)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for the dataset (index or character)}
}
\value{
A data frame
}
\description{
Get data from study
}
\examples{

s <- study() \%>\% add_data("test", data.frame(x = 1:10))
dat <- get_data(s, "test")

}
