% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_env_name}
\alias{get_env_name}
\title{Get Package}
\usage{
get_env_name(f)
}
\arguments{
\item{f}{the function to search (as a character string)}
}
\value{
a list of packages/environments the function is in
}
\description{
Get Package
}
\keyword{internal}
