% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_html}
\alias{get_html}
\title{Get result value as HTML}
\usage{
get_html(result = NULL, analysis_id = 1, digits = getOption("digits", 3))
}
\arguments{
\item{result}{The name of the result value to retrieve (omit to return all results)}

\item{analysis_id}{The id for the relevant analysis (index or character) defaults to the first analysis in the list}

\item{digits}{Number of digits to round to (defaults to the value for options("digits")}
}
\value{
html
}
\description{
Get result value as HTML
}
