% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_orcid.R
\name{get_orcid}
\alias{get_orcid}
\title{Get ORCiD from Name}
\usage{
get_orcid(family, given = "*")
}
\arguments{
\item{family}{The family (last) name to search for}

\item{given}{An optional given (first) name to search for. Initials will be converted from, e.g., L M to L\* M\*}
}
\value{
A vector of matching ORCiDs
}
\description{
Get ORCiD from Name
}
\examples{
get_orcid("DeBruine", "Lisa")

}
