% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_power.R
\name{get_power}
\alias{get_power}
\title{Get Power Calculations as List}
\usage{
get_power(study, values = FALSE)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{values}{Whether or not to retun criterion values (default FALSE)}
}
\value{
a list of power (and criteria) values
}
\description{
Get Power Calculations as List
}
\examples{
study() \%>\%
 add_hypothesis("H1") \%>\%
 add_analysis("A1", t.test(y~A, data = D1)) \%>\%
 add_criterion("C1", "p.value", "<", 0.05) \%>\%
 add_analysis("A2", t.test(y~A, data = D2)) \%>\%
 add_criterion("C2", "p.value", "<", 0.05) \%>\%
 add_eval("corroboration", "", "C1 & C2") \%>\%
 add_eval("falsification", "", "!C1 & !C2") \%>\%
 add_sim_data("D1", between = 2, n = 25, mu = c(0, 0.5)) \%>\%
 add_sim_data("D2", between = 2, n = 50, mu = c(0, 0.5)) \%>\%
 study_power(rep = 100) \%>\%
 get_power()

}
