% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_res_value}
\alias{get_res_value}
\title{Get value from results list}
\usage{
get_res_value(txt, results)
}
\arguments{
\item{txt}{text of result to check against names}

\item{results}{named list of results}
}
\value{
value from results list or the txt if not found
}
\description{
Get value from results list
}
\keyword{internal}
