% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEnvName}
\alias{getEnvName}
\title{Get Package Name}
\usage{
getEnvName(f)
}
\arguments{
\item{f}{the function to search (as a character string)}
}
\value{
a list of packages the function is in
}
\description{
Get Package Name
}
\keyword{internal}
