% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_nowt}
\alias{if_nowt}
\title{Replace values if NULL, NA or empty}
\usage{
if_nowt(x, replace = "", test_for = c("null", "na", "trim"))
}
\arguments{
\item{x}{vector or list to test}

\item{replace}{value to replace with}

\item{test_for}{values to test for ("null" replaces NULL values, "na", replaces NA values, "trim" replaces empty string after trimws()}
}
\value{
vector or list with replaced values
}
\description{
Replace values if NULL, NA or empty
}
\examples{
if_nowt(NULL)
if_nowt(NA)
if_nowt("   ")
if_nowt(c(1, 2, NA), replace = 0)
if_nowt(list(NULL, NA, " "))
if_nowt(list(NULL, NA, " "), test_for = "null")
if_nowt(list(NULL, NA, " "), test_for = "na")
if_nowt(list(NULL, NA, " "), test_for = "trim")
}
