% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_kin.R
\docType{data}
\name{kin}
\alias{kin}
\title{Kinship}
\format{
A data frame with 24 rows and 4 variables:
\describe{
\item{trust_self}{Number of trusting moves towards self-morphs}
\item{trust_other}{Number of trusting moves towards self-morphs}
\item{recip_self}{Number of reciprocating moves towards other-morphs}
\item{recip_other}{Number of reciprocating moves towards other-morphs}
}
}
\usage{
kin
}
\description{
Data from DeBruine (2002) Facial Resemblance Enhances Trust, PRSLB. https://osf.io/ewfhs/
}
\keyword{datasets}
