% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{load_params}
\alias{load_params}
\title{Load Params}
\usage{
load_params(params, study)
}
\arguments{
\item{params}{a list of parameter names and values}

\item{study}{the study object to get the data from}
}
\value{
params list
}
\description{
Load .data\link{id} and .data\link{id}$col references from the data
}
