% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Save data in files}
\usage{
make_data(study, dir = "data", format = c("tsv", "csv"), ...)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{dir}{The directory in which to save to data}

\item{format}{The data format to save with (defaults to tsv)}

\item{...}{Further arguments to pass to write.table or write.csv}
}
\value{
the study object
}
\description{
Save data in files
}
