% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_func}
\alias{make_func}
\title{Make a function}
\usage{
make_func(func, code, return = "", envir = .GlobalEnv)
}
\arguments{
\item{func}{the function name}

\item{code}{the function body}

\item{return}{a list of names of objects to return from the function (if blank, defaults to last value from the code, which should be a named list)}

\item{envir}{the environment in which to define the function}
}
\value{
creates a function
}
\description{
Make a function
}
\keyword{internal}
