% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis.R
\name{make_return}
\alias{make_return}
\title{Make return list}
\usage{
make_return(return)
}
\arguments{
\item{return}{a vector of names of objects to return from the function}
}
\value{
character string of return list
}
\description{
Make return list
}
\keyword{internal}
