% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_script.R
\name{make_script}
\alias{make_script}
\title{Make a script with all analyses}
\usage{
make_script(
  study,
  path = NULL,
  data_path = "data",
  data_format = "tsv",
  show_codebook = TRUE,
  use_rmarkdown = TRUE,
  header = TRUE
)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{path}{Path to write script to (outputs text if NULL)}

\item{data_path}{Where to load data from external files (you may need to creates data files with `make_data or edit the script for your data paths); loads from the script if set to NULL}

\item{data_format}{The data format to save with (defaults to tsv)}

\item{show_codebook}{Include codebook for each dataset in comments}

\item{use_rmarkdown}{Creates an Rmarkdown file if TRUE, a .R file if FALSE}

\item{header}{Whether the rmarkdown version should have a header}
}
\value{
Text of the script if path is NULL
}
\description{
Make a script with all analyses
}
\examples{
s <- study() \%>\%
  add_data("my_cars", mtcars) \%>\%
  add_analysis("A1", cor.test(my_cars$mpg, my_cars$wt)) \%>\%
  study_analyse()

# get Rmd text output
make_script(s) \%>\% cat()
# get R text output
make_script(s, use_rmarkdown = FALSE) \%>\% cat()

}
