% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlist}
\alias{nlist}
\title{Make a Named List}
\usage{
nlist(names, values)
}
\arguments{
\item{names}{a vector of the names}

\item{values}{a vector of the values}
}
\value{
a named list
}
\description{
Make a Named List
}
\examples{
n <- LETTERS[1:3]
val <- c("cat", "dog", "ferret")
nlist(n, val)

}
