% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_analyses}
\alias{output_analyses}
\title{Output analyses}
\usage{
output_analyses(study, header_lvl = 2)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{header_lvl}{The starting header level for the section (defaults to 2)}
}
\value{
The study object
}
\description{
Output analysis plan specified in the json file
}
