% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_custom_code}
\alias{output_custom_code}
\title{Output custom code}
\usage{
output_custom_code(study, analysis_id = 1)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{analysis_id}{The id or index for the analysis code to output (defaults to index 1)}
}
\value{
string of the function definition
}
\description{
Output custom code specified in an analysis component
}
\examples{

custom <- function() { 1:10 }
s <- study() \%>\%
  add_analysis("custom", t.test(rnorm(100))) \%>\%
  output_custom_code()
}
