% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_info}
\alias{output_info}
\title{Output study info}
\usage{
output_info(study, header_lvl = 2, output = c("md", "html", "text"))
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{header_lvl}{The starting header level for the section (defaults to 2)}

\item{output}{whether the output should be markdown, html, or plain text (defaults to md)}
}
\value{
character string with a human-readable summary of the study info
}
\description{
Output study info specified in the json file
}
