% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.scivrs_author}
\alias{print.scivrs_author}
\title{Print Author List}
\usage{
\method{print}{scivrs_author}(x, ...)
}
\arguments{
\item{x}{The scivrs_author list}

\item{...}{Additional parameters for print}
}
\description{
Print Author List
}
\keyword{internal}
