% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.scivrs_authors}
\alias{print.scivrs_authors}
\title{Print Authors List}
\usage{
\method{print}{scivrs_authors}(x, ...)
}
\arguments{
\item{x}{The scivrs_authors list}

\item{...}{Additional parameters for print}
}
\description{
Print Authors List
}
