% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.scivrs_codebook}
\alias{print.scivrs_codebook}
\title{Print Codebook Object}
\usage{
\method{print}{scivrs_codebook}(x, ...)
}
\arguments{
\item{x}{The scivrs_codebook list}

\item{...}{Additional parameters for print}
}
\description{
Print Codebook Object
}
