% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.scivrs_results}
\alias{print.scivrs_results}
\title{Print Results List}
\usage{
\method{print}{scivrs_results}(x, ...)
}
\arguments{
\item{x}{The scivrs_results list}

\item{...}{Additional parameters for print}
}
\description{
Print Results List
}
