% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.scivrs_study}
\alias{print.scivrs_study}
\title{Print Study Object}
\usage{
\method{print}{scivrs_study}(x, ...)
}
\arguments{
\item{x}{The scivrs_study list}

\item{...}{Additional parameters for print}
}
\description{
Print Study Object
}
