% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scienceverse-package.R
\docType{package}
\name{scienceverse-package}
\alias{scienceverse}
\alias{scienceverse-package}
\title{scienceverse: Machine-Readable Study Descriptions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Generate and process machine-readable study descriptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/scienceverse/scienceverse}
  \item Report bugs at \url{https://github.com/scienceverse/scienceverse/issues}
}

}
\author{
\strong{Maintainer}: Lisa DeBruine \email{debruine@gmail.com} (\href{https://orcid.org/0000-0002-7523-5539}{ORCID})

Authors:
\itemize{
  \item Daniel Lakens \email{lakens@gmail.com} (\href{https://orcid.org/0000-0002-0247-239X}{ORCID})
}

Other contributors:
\itemize{
  \item Brenton Wiernik (\href{https://orcid.org/0000-0001-9560-6336}{ORCID}) [contributor]
}

}
\keyword{internal}
