% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scienceverse_options.R
\name{scienceverse_options}
\alias{scienceverse_options}
\title{Set/get global scienceverse options}
\usage{
scienceverse_options(...)
}
\arguments{
\item{...}{One of four: (1) nothing, then returns all options as a list; (2) a name of an option element, then returns its value; (3) a name-value pair which sets the corresponding option to the new value (and returns nothing), (4) a list with option-value pairs which sets all the corresponding arguments.}
}
\value{
a list of options, values of an option, or nothing
}
\description{
Global options are used to set default values.
}
\examples{

scienceverse_options() # see all options

scienceverse_options("quiet") # see value of scienceverse.quiet
\dontrun{
scienceverse_options(quiet = TRUE)
}
}
