% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scivrs_app.R
\name{scivrs_app}
\alias{scivrs_app}
\title{Launch Shiny App}
\usage{
scivrs_app(...)
}
\arguments{
\item{...}{arguments to pass to shiny::runApp}
}
\description{
Create a meta-study file interactively in a shiny app that runs locally in RStudio or your web browser (recommended). It does not connect to the web at all, so your data are completely private.
}
\examples{
\dontrun{ scivrs_app() }

}
