% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\title{Create a study object}
\usage{
study(name = "Demo Study", ...)
}
\arguments{
\item{name}{The name of the study or a file path to a JSON meta-study file}

\item{...}{further arguments to add}
}
\value{
A study object with class scivrs_study
}
\description{
Create a new study object or load a study from the JSON meta-study file
}
\examples{

s <- study("Demo Study")
study_to_json(s)

}
