% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_analyse}
\alias{study_analyse}
\title{Run analysis}
\usage{
study_analyse(study)
}
\arguments{
\item{study}{A study list object with class reg_study}
}
\value{
A study object with class reg_study
}
\description{
Run the analyses on the data
}
