% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_study_demo.R
\docType{data}
\name{study_demo}
\alias{study_demo}
\title{Study Demo}
\format{
A list with 7 items:
\describe{
\item{name}{Study name}
\item{info}{General info about the study}
\item{authors}{Info about the authors}
\item{hypotheses}{Info about the hypotheses}
\item{methods}{Info about the methods (currently blank)}
\item{data}{Info about the data}
\item{analyses}{Info about the analyses}
}
}
\usage{
study_demo
}
\description{
A demo scienceverse study object for DeBruine (2002) Facial Resemblance Enhances Trust, PRSLB.
}
\keyword{datasets}
