% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_eval.R
\name{study_eval}
\alias{study_eval}
\title{Evaluate study results}
\usage{
study_eval(study)
}
\arguments{
\item{study}{A study list object with class scivrs_study}
}
\value{
A study object with class scivrs_study
}
\description{
Evaluate study results
}
