% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_from_json.R
\name{study_from_json}
\alias{study_from_json}
\title{Get study from json file}
\usage{
study_from_json(filename)
}
\arguments{
\item{filename}{The name of the scienceverse-formatted json file}
}
\value{
A study object with class reg_study
}
\description{
Get study from json file
}
