% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_from_xml.R
\name{study_from_xml}
\alias{study_from_xml}
\title{Get study from XML file}
\usage{
study_from_xml(filename, xml_type = c("auto", "grobid"))
}
\arguments{
\item{filename}{the path to the XML file}

\item{xml_type}{the type of xml file to parse}
}
\value{
A study object with class scivrs_study
}
\description{
Currently only works for XML created by grobid
}
\examples{
grobid_dir <- system.file("grobid", package="scienceverse")
filename <- file.path(grobid_dir, "eyecolor.pdf.tei.xml")
study <- study_from_xml(filename)

}
