% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_power.R
\name{study_power}
\alias{study_power}
\title{Power Analysis through Simulation}
\usage{
study_power(study, rep = 100)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{rep}{The number of simulations to run}
}
\value{
A study object with class reg_study
}
\description{
This function is experimental. Check power analyses with an external package before using for important decisions.
}
\examples{

study <- study() \%>\%
 add_hypothesis("H1") \%>\%
 add_analysis("A1", t.test(y~A, data = D1)) \%>\%
 add_criterion("C1", "p.value", "<", 0.05) \%>\%
 add_analysis("A2", t.test(y~A, data = D2)) \%>\%
 add_criterion("C2", "p.value", "<", 0.05) \%>\%
 add_eval("corroboration", "", "C1 & C2") \%>\%
 add_eval("falsification", "", "!C1 & !C2") \%>\%
 add_sim_data("D1", between = 2, n = 25, mu = c(0, 0.5)) \%>\%
 add_sim_data("D2", between = 2, n = 50, mu = c(0, 0.5)) \%>\%
 study_power(rep = 100)

}
