% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_report}
\alias{study_report}
\title{Generate a Report}
\usage{
study_report(study, type = "prereg")
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{type}{The type of report c("prereg", "postreg")}
}
\value{
A study object with class reg_study
}
\description{
Generate a study report
}
\examples{

mystudy <- study("Iris Study") \%>\%
  add_hypothesis(1, "Petal length and width will be significantly correlated") \%>\%
  add_analysis(1, "cor.test", list(
    x = ".data$Petal.Length",
    y = ".data$Petal.Width"
  )) \%>\%
  add_criterion(1, 1, "p.value", "<", 0.05) \%>\%
  add_data(1, iris) \%>\%
  study_analyse() \%>\%
  study_report(type = "postreg")

}
