% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_to_json.R
\name{study_to_json}
\alias{study_to_json}
\title{Study Object to JSON string}
\usage{
study_to_json(study, data_values = TRUE)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{data_values}{Whether to include data values in the JSON file (defaults to TRUE)}
}
\value{
A prettified JSON string
}
\description{
Convert a study object to a JSON string
}
\examples{

study() \%>\% study_to_json()

}
