% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test.default}
\alias{t_test.default}
\title{T-test with N}
\usage{
\method{t_test}{default}(
  x,
  y = NULL,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  paired = FALSE,
  var.equal = FALSE,
  conf.level = 0.95,
  ...
)
}
\description{
See \code{stats::\link[stats]{t.test}} for details.
}
