% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{T-test with N (formula)}
\usage{
t_test(x, ...)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{...}{further arguments to be passed to or from methods.}
}
\description{
See \code{stats::\link[stats]{t.test}} for details.
}
