% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test_n}
\alias{t_test_n}
\title{Calculate Ns for t_test}
\usage{
t_test_n(x, y = NULL, paired = FALSE, names = NULL)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{n optional (non-empty) numeric vector of data values.}

\item{paired}{a logical indicating whether you want a paired t-test.}

\item{names}{optional group names for 2-sample tests.}
}
\value{
list of Ns
}
\description{
Calculate Ns for t_test
}
