% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis.R
\name{update_analysis}
\alias{update_analysis}
\title{Update an analysis}
\usage{
update_analysis(
  study,
  id,
  code = NULL,
  return = "",
  type = c("func", "text", "file"),
  new_id = NULL,
  ...
)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for this analysis (index or character) if an analysis with this id already exists, it will overwrite it}

\item{code}{The code to run or a file name containing the code}

\item{return}{A list of object names to return from the code}

\item{type}{Whether the code is a function, text, or file name}

\item{new_id}{A new (character) ID for this hypothesis}

\item{...}{further arguments to add}
}
\value{
A study object with class scivrs_study
}
\description{
Update an analysis
}
\examples{
s <- study() \%>\% add_analysis("A1", rnorm(10))
s <- update_analysis(s, 1, code = rnorm(20), new_id = "A1a")
s
}
