% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_criterion.R
\name{update_criterion}
\alias{update_criterion}
\title{Update a criterion}
\usage{
update_criterion(
  study,
  id,
  result = NULL,
  operator = NULL,
  comparator = NULL,
  analysis_id = NULL,
  hypothesis_id = NULL,
  new_id = NULL
)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{A unique name to refer to the criterion with in the evaluation}

\item{result}{The name of the item in the analysis results list to compare}

\item{operator}{The operator for comparison c("<", "=", ">", "!=")}

\item{comparator}{The value to compare}

\item{analysis_id}{The id for the relevant analysis (index or character) if NULL, assigns to the last analysis in the list}

\item{hypothesis_id}{The id for the hypothesis (index or character) if NULL, assigns to the last hypothesis in the list}

\item{new_id}{A new (character) ID for this criterion}
}
\value{
A study object with class scivrs_study
}
\description{
Update a criterion
}
\examples{
s <- study() \%>\%
  add_hypothesis("H1") \%>\%
  add_analysis("A1", t.test(rnorm(100))) \%>\%
  add_criterion("p", "p.value", ">", 0.05) \%>\%
  study_analyse()

s <- update_criterion(s, "p", operator = "<",
                      analysis_id = "A1",
                      hypothesis_id = "H1" ) \%>\%
     study_analyse()

}
