% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hypothesis.R
\name{update_hypothesis}
\alias{update_hypothesis}
\title{Update a hypothesis}
\usage{
update_hypothesis(study, id = 1, description = NULL, new_id = NULL)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for this hypothesis (index or character)}

\item{description}{The text description of the hypothesis}

\item{new_id}{A new (character) ID for this hypothesis}
}
\value{
A study object with class scivrs_study
}
\description{
Update a hypothesis
}
\examples{
s <- study() \%>\% add_hypothesis("H1", "My hypothesis")
s <- update_hypothesis(s, 1, "Better description", "H1a")
s
}
