% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_method.R
\name{update_method}
\alias{update_method}
\title{Update a method}
\usage{
update_method(study, id = 1, description = NULL, new_id = NULL)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for this method (index or character)}

\item{description}{The text description of the method}

\item{new_id}{A new (character) ID for this method}
}
\value{
A study object with class scivrs_study
}
\description{
Update a method
}
\examples{
s <- study() \%>\% add_method("M1", "My method")
s <- update_method(s, 1, "Better description", "M1a")
s
}
