% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis.R
\name{add_analysis}
\alias{add_analysis}
\title{Add Analysis}
\usage{
add_analysis(study, id = NULL, code = "", return = "", ...)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for this analysis (index or character) if an analysis with this id already exists, it will overwrite it}

\item{code}{The code to run or a file name containing the code}

\item{return}{A list of object names to return from the code}

\item{...}{further arguments to add}
}
\value{
A study object with class scivrs_study
}
\description{
Add an analysis to a study object
}
\examples{

s <- study() \%>\%
  add_hypothesis("H1", "Petal width and length will be positively correlated.") \%>\%
  add_analysis("A1", cor.test(dat$Petal.Width, dat$Petal.Length))
study_to_json(s)

}
