% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{add_analysis}
\alias{add_analysis}
\title{Add Analysis}
\usage{
add_analysis(study, id = NULL, func = "list", params = list(),
  code = NULL)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{id}{The id for this analysis (index or character) if NULL, this creates a new analysis, if an analysis with this id already exists, it will overwrite it}

\item{func}{The name of the function to run}

\item{params}{A list of parameters for the function arguments}

\item{code}{Code to define custom functions}
}
\value{
A study object with class reg_study
}
\description{
Add an analysis to a study object
}
\examples{

mystudy <- study("Iris Petals") \%>\%
  add_hypothesis(1, "Petal length and width will be significantly correlated") \%>\%
  add_criterion(1, 1, "p.value", "<", 0.05) \%>\%
  add_analysis(1, "cor.test", list(
    x = ".data[1]$Petal.Length",
    y = ".data[1]$Petal.Width",
    alternative = "two.sided",
    method = "pearson",
    conf.level = 0.95
  ))

}
