% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_criterion.R
\name{add_criterion}
\alias{add_criterion}
\title{Add Criterion}
\usage{
add_criterion(
  study,
  id,
  result,
  operator,
  comparator,
  hypothesis_id = NULL,
  analysis_id = NULL
)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{A unique name to refer to the criterion with in the evaluation}

\item{result}{The name of the item in the analysis results list to compare}

\item{operator}{The operator for comparison c("<", "=", ">", "!=")}

\item{comparator}{The value to compare}

\item{hypothesis_id}{The id for the hypothesis (index or character) if NULL, assigns to the last hypothesis in the list}

\item{analysis_id}{The id for the relevant analysis (index or character) if NULL, assigns to the last analysis in the list}
}
\value{
A study object with class scivrs_study
}
\description{
Add a criterion to a hypothesis in a study object
}
\examples{

s <- study() \%>\%
  add_hypothesis("H1", "Petal width and length will be positively correlated.") \%>\%
  add_analysis("A1", cor.test(dat$Petal.Width, dat$Petal.Length)) \%>\%
  add_criterion("sig", "p.value", "<", 0.05) \%>\%
  add_criterion("pos", "estimate", ">", 0)

}
