% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{add_data}
\alias{add_data}
\title{Add Data}
\usage{
add_data(study, data = NULL, id = NULL)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{data}{The dataset as a data.frame, codebook object, path to a data file, or path to a codebook file}

\item{id}{The id for this dataset (index or character) if NULL, this creates a new dataset, if an analysis with this id already exists, it will overwrite it}
}
\value{
A study object with class reg_study
}
\description{
Add a dataset to a study object
}
\examples{

mystudy <- study() \%>\%
  add_data(iris)

}
