% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Add Data}
\usage{
add_data(study, id, data = NULL, coldesc = NULL)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{id}{The id for this dataset (index or character) if a dataset with this id already exists, it will overwrite it}

\item{data}{The dataset as a data.frame, codebook object, path to a data file, or path to a codebook file}

\item{coldesc}{Optional named list of column descriptions}
}
\value{
A study object with class reg_study
}
\description{
Add a dataset to a study object
}
\examples{

s <- study() \%>\% add_data("dat", iris)
study_json(s)

}
