% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{add_data}
\alias{add_data}
\title{Add Data}
\usage{
add_data(study, id, data = NULL, vardesc = NULL, design = NULL, ...)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for this dataset (index or character) if a dataset with this id already exists, it will overwrite it}

\item{data}{The dataset as a data.frame, codebook object, path to a data file, or path to a codebook file}

\item{vardesc}{Optional variable properties (see \code{codebook})}

\item{design}{A faux design specification}

\item{...}{Further dataset properties (see \code{codebook})}
}
\value{
A study object with class scivrs_study
}
\description{
Add a dataset to a study object
}
\examples{

s <- study() \%>\% add_data("dat", iris)
study_to_json(s)

}
