% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_eval.R
\name{add_eval}
\alias{add_eval}
\title{Add an evaluation}
\usage{
add_eval(study, type, evaluation, description = "", hypothesis_id = NULL)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{type}{"corroboration" or "falsification" (or "c"/"f")}

\item{evaluation}{A logical representation of these conditions using the criteria IDs, parentheses, &, | and ! (e.g., "(c1 & c2) | (c3 & !c4)")}

\item{description}{A verbal description of the conditions for corroborating the hypothesis}

\item{hypothesis_id}{The id for the hypothesis (index or character) if NULL, assigns to the last hypothesis in the list}
}
\value{
A study object with class scivrs_study
}
\description{
Add an evaluation criterion to a hypothesis in a study object
}
\examples{

s <- study() \%>\%
  add_hypothesis("H1", "Petal width and length will be positively correlated.") \%>\%
  add_analysis("A1", cor.test(dat$Petal.Width, dat$Petal.Length)) \%>\%
  add_criterion("sig", "p.value", "<", 0.05) \%>\%
  add_criterion("pos", "estimate", ">", 0) \%>\%
  add_eval("corroboration", "sig & pos",
           "Petal width is significantly and positively correlated to length"
           ) \%>\%
  add_eval("falsification", "sig & !pos",
           "Petal width is significantly and negatively correlated to length"
           )
s
}
