% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hypothesis.R
\name{add_hypothesis}
\alias{add_hypothesis}
\title{Add a hypothesis}
\usage{
add_hypothesis(study, id = NULL, description = "Describe your hypothesis", ...)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{id}{The id for this hypothesis (index or character) if a hypothesis with this id already exists, it will overwrite it}

\item{description}{The text description of the hypothesis}

\item{...}{further arguments to add}
}
\value{
A study object with class scivrs_study
}
\description{
Add a hypothesis to a study object
}
\examples{

s <- study() \%>\%
  add_hypothesis("H1", "Petal width and length will be positively correlated.")
study_to_json(s)

}
