% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{add_hypothesis}
\alias{add_hypothesis}
\title{Add Hypothesis}
\usage{
add_hypothesis(study, description = "Describe your hypothesis",
  evaluation = "&", id = NULL)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{description}{The text descriiption of the hypothesis}

\item{evaluation}{The rule for evaluating this criterion c("&", "and", "|", "or")}

\item{id}{The id for this hypothesis (index or character) if NULL, this creates a new hypothesis, if a hypothesis with this id already exists, it will overwrite it}
}
\value{
A study object with class reg_study
}
\description{
Add a hypothesis to a study object
}
\examples{

mystudy <- study("Iris Petals") \%>\%
  add_hypothesis("Petal length and width will be significantly correlated")

}
