% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_author.R
\name{author}
\alias{author}
\title{Author in CRediT Format}
\usage{
author(surname, given = "", orcid = NULL, roles = c(), ...)
}
\arguments{
\item{surname}{a character string with the author's last name(s)}

\item{given}{a character string with the author's given name(s)}

\item{orcid}{the author's unique ORCiD (see https://orcid.org/)}

\item{roles}{a vector of roles from the CRediT taxonomy (see https://casrai.org/credit/); use credit_roles() to view the full list}

\item{...}{further info to add to author object}
}
\value{
a list with class scivrs_author
}
\description{
Author in CRediT Format
}
\examples{
author(
  surname = "DeBruine", given = "Lisa M.",
  orcid = "0000-0002-7523-5539",
  roles = c("Conceptualization", "Methodology")
)
}
