% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_jats.R
\name{author_jats}
\alias{author_jats}
\title{Convert scivrs_author list to CRediT JATS format}
\usage{
author_jats(author)
}
\arguments{
\item{author}{list with class scivrs_author or list with class reg_study and at least one entry for authors}
}
\value{
string in JATS format (see https://jats4r.org/credit-taxonomy)
}
\description{
Convert scivrs_author list to CRediT JATS format
}
\examples{
ld <- author(
  orcid = "0000-0002-7523-5539",
  surname = "DeBruine", given = "Lisa M.",
  roles = list("Conceptualization", "Methodology")
)
author_jats(ld)
}
