% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Create PsychDS Codebook from Data}
\usage{
codebook(data, coldesc = NULL, as_json = TRUE)
}
\arguments{
\item{data}{The data frame to generate a codebook for}

\item{coldesc}{Optional named list of column descriptions}

\item{as_json}{Whether the output should be a list or JSON format}
}
\value{
a list or json-formatted codebook
}
\description{
Create PsychDS Codebook from Data
}
\examples{

codebook(iris)

}
