% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_result.R
\name{get_result}
\alias{get_result}
\title{Get Result Value}
\usage{
get_result(study, result = NULL, analysis_id = 1)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{result}{The name of the result value to retrieve (omit to return all results)}

\item{analysis_id}{The id for the relevant analysis (index or character) defaults to the first analysis in the list}
}
\value{
value of names result or list of results
}
\description{
Get Result Value
}
\examples{
s <- study() \%>\%
  add_hypothesis("H1", "Petals are wider than long") \%>\%
  add_analysis("A1", t.test(iris$Petal.Width, iris$Petal.Height)) \%>\%
  add_criterion("p", "p.value", "<", .05) \%>\%
  study_analyse()

 get_result(s, "p.value", "A1")

}
