% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nowt}
\alias{is_nowt}
\title{Check if values are NULL, NA or empty}
\usage{
is_nowt(x, test_for = c("null", "na", "trim", "empty"))
}
\arguments{
\item{x}{vector or list to test}

\item{test_for}{values to test for ("null" replaces NULL values, "na", replaces NA values, "trim" replaces empty string after trimws(), "empty" replaces empty lists)}
}
\value{
vector or list of logical values
}
\description{
Check if values are NULL, NA or empty
}
\examples{
x <- list(NULL, NA, " ", list())
is_nowt(x)
is_nowt(x, test_for = "null")
is_nowt(x, test_for = "na")
is_nowt(x, test_for = "trim")
is_nowt(x, test_for = "empty")

}
