% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_func}
\alias{make_func}
\title{Make a function}
\usage{
make_func(func, args, code, return = ".Last.value", envir = .GlobalEnv)
}
\arguments{
\item{func}{the function name}

\item{args}{the function arguments}

\item{code}{the function body}

\item{return}{a list of names of objects to return from the function}

\item{envir}{the environment in which to define the function}
}
\value{
creates a function
}
\description{
Make a function
}
\keyword{internal}
