% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis.R
\name{make_func}
\alias{make_func}
\title{Make a function}
\usage{
make_func(func, code, envir = .GlobalEnv)
}
\arguments{
\item{func}{the function name}

\item{code}{the function body (as character, vector, or list)}

\item{envir}{the environment in which to define the function}
}
\value{
creates a function
}
\description{
Make a function
}
\keyword{internal}
