% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlist}
\alias{nlist}
\title{Make a Named List}
\usage{
nlist(names, values)
}
\arguments{
\item{names}{a vector or list of the names}

\item{values}{a vector or list of the values}
}
\value{
a named list
}
\description{
Convenience function for making a named list from a vector or list of names and a vector or list of values
}
\examples{
n <- LETTERS[1:3]
val <- c("cat", "dog", "ferret")
nlist(n, val)

nlist(list("A", "B"), 1:2)

}
