% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{output_analyses}
\alias{output_analyses}
\title{Output analyses}
\usage{
output_analyses(study)
}
\arguments{
\item{study}{A study list object with class reg_study}
}
\value{
The study object
}
\description{
Output analysis plan specified in the json file
}
