% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_hypotheses}
\alias{output_hypotheses}
\title{Output hypotheses}
\usage{
output_hypotheses(study)
}
\arguments{
\item{study}{A study list object with class scivrs_study}
}
\value{
The study object
}
\description{
Output hypotheses specified in the json file
}
