% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_hypotheses}
\alias{output_hypotheses}
\title{Output hypotheses}
\usage{
output_hypotheses(study, header_lvl = 2, output = c("md", "html", "text"))
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{header_lvl}{The starting header level for the section (defaults to 2)}

\item{output}{whether the output should be markdow, html, or plain text (defaults to md)}
}
\value{
character string with a human-readable summary of the hypotheses
}
\description{
Output hypotheses specified in the json file
}
