% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_results}
\alias{output_results}
\title{Output results}
\usage{
output_results(
  study,
  header_lvl = 2,
  output = c("md", "html", "text"),
  digits = 3
)
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{header_lvl}{The starting header level for the section (defaults to 2)}

\item{output}{whether the output should be markdow, html, or plain text (defaults to md)}

\item{digits}{integer indicating the number of decimal places.}
}
\value{
character string with a human-readable summary of the results
}
\description{
Output results specified in the json file
}
