% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{output_results}
\alias{output_results}
\title{Output results}
\usage{
output_results(study, digits = 3)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{digits}{integer indicating the number of decimal places.}
}
\value{
The study object
}
\description{
Output results specified in the json file
}
