% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study}
\alias{study}
\title{Study}
\usage{
study(name = "Demo Study")
}
\arguments{
\item{name}{The name of the study or a file path to a json file}
}
\value{
A study object with class reg_study
}
\description{
Create or load a study object
}
\examples{

mystudy <- study("Iris Petals")

}
