% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_analyse}
\alias{study_analyse}
\alias{study_analyze}
\title{Run analysis}
\usage{
study_analyse(study)

study_analyze(study)
}
\arguments{
\item{study}{A study list object with class reg_study}
}
\value{
A study object with class reg_study
}
\description{
Run the analyses on the data
}
\examples{
study() \%>\%
  add_hypothesis() \%>\%
  add_analysis("cor.test", list(
    x = ".data[1]$Petal.Width",
    y = ".data[1]$Petal.Length"
  )) \%>\%
  add_criterion(
    result = "p.value",
    operator = "<",
    comparator = 0.05
  ) \%>\%
  add_data(iris) \%>\%
  study_analyse()

}
