% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_report.R
\name{study_report}
\alias{study_report}
\title{Generate a Report}
\usage{
study_report(study, template = "prereg", filename = "study.html")
}
\arguments{
\item{study}{A study list object with class scivrs_study}

\item{template}{The type of report c("prereg", "postreg") or a path to a custom template}

\item{filename}{The file path to save to}
}
\value{
A study object with class scivrs_study
}
\description{
Generate a study report (deprecated, use study_save)
}
