% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_report.R
\name{study_report}
\alias{study_report}
\title{Generate a Report}
\usage{
study_report(study, template = "prereg", filename = "study.html")
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{template}{The type of report c("prereg", "postreg") or a path to a custom template}

\item{filename}{The file path to save to}
}
\value{
A study object with class reg_study
}
\description{
Generate a study report
}
\examples{

\dontrun{
study() \%>\%
  add_hypothesis("H1", "Petal width and length will be positively correlated.") \%>\%
  add_analysis("A1", cor.test(dat$Petal.Width, dat$Petal.Length)) \%>\%
  add_criterion("sig", "p.value", "<", 0.05) \%>\%
  add_criterion("pos", "estimate", ">", 0) \%>\%
  add_eval("corroboration",
           "Petal width is significantly and positively correlated to length",
           "sig & pos") \%>\%
  add_eval("falsification",
           "Petal width is significantly and negatively correlated to length",
           "sig & !pos") \%>\%
  add_data("dat", iris) \%>\%
  study_analyse() \%>\%
  study_report(template = "postreg")
}

}
