% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_report}
\alias{study_report}
\title{Generate a Report}
\usage{
study_report(study, template = "prereg", filename = "study.html")
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{template}{The type of report c("prereg", "postreg") or a path to a custom template}

\item{filename}{The file path to save to}
}
\value{
A study object with class reg_study
}
\description{
Generate a study report
}
\examples{

\dontrun{
mystudy <- study("Iris Study") \%>\%
  add_hypothesis("Petal length and width will be significantly correlated") \%>\%
  add_analysis("cor.test", list(
    x = ".data[1]$Petal.Length",
    y = ".data[1]$Petal.Width"
  )) \%>\%
  add_criterion("p.value", "<", 0.05) \%>\%
  add_data(iris) \%>\%
  study_analyse() \%>\%
  study_report(template = "postreg")
}

}
