% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{study_save}
\alias{study_save}
\title{Save study}
\usage{
study_save(study, filename = "study.json", data_values = TRUE)
}
\arguments{
\item{study}{A study list object with class reg_study}

\item{filename}{The name to save the file}

\item{data_values}{Whether to include data values in the JSON file (defaults to TRUE)}
}
\value{
A study object with class reg_study
}
\description{
Save the study framework to a JSON file
}
