% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test.formula}
\alias{t_test.formula}
\title{T-test with N (formula)}
\usage{
\method{t_test}{formula}(formula, data, subset, na.action, ...)
}
\description{
See \code{stats::\link[stats]{t.test}} for details.
}
